+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Improve switch speed when changing profiles
+* Allow to switch profile when sharing or forwarding
 * Don't notify notification-to-all from in-chat apps if the chat is muted
 * Allow to see inbox quota for all relays in connectivity screen
 * Truncate file names in the middle, not at the end; important information are more often at the end